requirements.txt
numpy==1.16.0
torch==0.4.1
gym==0.15.3
PyYAML==5.3
tensorboardX==2.0
sacred==0.7.4
